% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta_function.R
\name{theta}
\alias{theta}
\title{Theta and zeta indices for phylogenetic signal}
\usage{
theta(cat, tree, matching = "complete")
}
\arguments{
\item{cat}{A dataframe containing data for the categorical trait. The data preferably should be in different columns for each class of the trait and 0 indicating the species does not have this class, while 1 that it does. A species could have more than one categories.}

\item{tree}{The phylogenetic tree in phylo format. The tree must include all species in the trait dataset}

\item{matching}{The similarity criterion to be used "complete" for assuming two species as identical if they match in all classes, and "partial" if we assume that species are considered similar if they share at least one class of the trait}
}
\value{
The values of the theta and zeta indices. The values are between 0 and 1, and the smaller the value the stronger the signal
}
\description{
Theta and zeta indices for phylogenetic signal
}
\examples{
tree <- rtree(50)
cat <- matrix(sample(c(0,1), 200, replace=TRUE), ncol=4, nrow=50)
cat <-data.frame(cat)
rownames(cat)<-tree$tip.label
theta(cat, tree, "complete")
theta(cat, tree, "partial")
}
